socialnet/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
crispy-forms-semantic-ui==0.2.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.0.3
django-allauth==0.48.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==13.2.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
msgpack==1.0.3
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
pytest-asyncio==0.18.2
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wsproto==1.0.0
zope.interface==5.4.0
